Fix appshell navbar and aside widths for all routes
